requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.5.1
python-dotenv==1.0.0
h11==0.16.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt>=1.6.1
psycopg2-binary>=2.9
python-engineio==4.12.3
python-socketio==5.14.2
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
